+++ addons/gamification/tests/__init__.py
@@ -2,5 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import test_challenge
-from . import test_recompute_rank
 from . import test_karma_tracking
